+++ content/rules/panther_modifications_outside_cicd.py
